Move redirects functionality to `src/lib`
